[package] Add package management library. Fixes JB#39099
This adds nemoosupdate-packagemanagement and
nemoosupdate-packagemanagement-devel packages, containing a package
management API. The API provides a Package::Manager class that can be
subclassed to implement a package manager. The library includes a
package manager implemented using PackageKit, via the PackageKitManager
class.

In moving this source code to this library, the following changes have
been made:
- namespace of PackageManager
- renaming of ManagerPkgKit -> PackageKitManager
- publicly exporting Manager, Transaction, Collection and PackageKitManager
- adding of internal d-pointers for the publicly exported classes
- adding of methods to PackageKitManager that allow configuration of repos