[osupdate] Add OS update dbus client and service classes. Contributes to JB#39100
This adds abstract base class called OsUpdater that is
implemented by the OsUpdaterClient and OsUpdaterService. The
OsUpdaterClient is a wrapper for the os update dbus interface
whereas the OsUpdaterService is a dbus adapter that should
be inherited by a platform specific os updater to implement platform
specific checking if an os update is available, download an update,
and install the downloaded update.

DBus service name is "org.nemo.osupdater".